--- a/dot_shell_common/core/reporter.sh
+++ b/dot_shell_common/core/reporter.sh
@@ -199,30 +199,219 @@ print_test_summary() {
     fi
 }
 
-# Print actionable recommendations
+# Print actionable recommendations based on missing components
 print_recommendations() {
     local failed="$1"
     local adapter="$2"
     
     if [ $failed -gt 0 ] && [ $REPORT_LEVEL -ge $REPORT_NORMAL ]; then
-        printf "\n${BOLD}Recommendations${NC}\n"
+        printf "\n${BOLD}Actionable Recommendations${NC}\n"
         printf "${GRAY}────────────────────────────────────────${NC}\n"
         
+        # Parse failed tests to provide specific recommendations
+        generate_specific_recommendations "$adapter"
+        
+        # General recommendations based on adapter mode
         case "$adapter" in
             pre_apply)
-                printf "${YELLOW}Before running 'chezmoi apply':${NC}\n"
-                printf "  • Ensure all configuration files are present\n"
-                printf "  • Verify shell compatibility\n"
-                printf "  • Run tests with -v for more details\n"
+                printf "\n${YELLOW}Before running 'chezmoi apply':${NC}\n"
+                printf "  • Review specific recommendations above\n"
+                printf "  • Install missing required components (priority: required > recommended > optional)\n"
+                printf "  • Run tests with -v for detailed installation hints\n"
+                printf "  • Focus on fixing 'required' failures first\n"
                 ;;
             post_apply)
-                printf "${YELLOW}To fix configuration issues:${NC}\n"
-                printf "  • Run 'chezmoi apply' to update configuration\n"
-                printf "  • Check shell configuration files\n"
-                printf "  • Verify environment variables\n"
-                printf "  • Restart shell sessions after changes\n"
+                printf "\n${YELLOW}To fix configuration issues:${NC}\n"
+                printf "  • Install missing components using provided commands\n"
+                printf "  • Run 'chezmoi apply' to update configuration after installations\n"
+                printf "  • Check shell configuration files and restart shell sessions\n"
+                printf "  • Verify environment variables are properly set\n"
                 ;;
         esac
+    fi
+}
+
+# Generate specific recommendations based on failed test analysis
+generate_specific_recommendations() {
+    local adapter="$1"
+    local critical_missing=""
+    local recommended_missing=""
+    local git_issues=""
+    local shell_issues=""
+    
+    # Analyze test results to identify specific issues
+    printf "${BLUE}${ICON_INFO}${NC} Specific Issues Found:\n"
+    
+    # Parse TEST_RESULTS for failed tests with install hints
+    local old_ifs="$IFS"
+    IFS='
+'
+    for result_line in $TEST_RESULTS; do
+        IFS='|'
+        set -- $result_line
+        local category="$1"
+        local name="$2"
+        local test_status="$3"
+        local details="$4"
+        local priority="$5"
+        local install_hint="$6"
+        IFS="$old_ifs"
+        
+        if [ "$test_status" = "FAIL" ] && [ -n "$install_hint" ]; then
+            case "$priority" in
+                "required")
+                    printf "  ${RED}🔴 CRITICAL${NC}: %s\n" "$name"
+                    printf "     ${GRAY}Fix: %s${NC}\n" "$install_hint"
+                    critical_missing="${critical_missing}\n    - $name: $install_hint"
+                    ;;
+                "recommended")
+                    printf "  ${YELLOW}🟡 RECOMMENDED${NC}: %s\n" "$name"
+                    printf "     ${GRAY}Install: %s${NC}\n" "$install_hint"
+                    recommended_missing="${recommended_missing}\n    - $name: $install_hint"
+                    ;;
+                "optional")
+                    printf "  ${BLUE}🔵 OPTIONAL${NC}: %s (can skip)\n" "$name"
+                    ;;
+            esac
+            
+            # Categorize issues for specific advice
+            case "$category" in
+                "config")
+                    if echo "$name" | grep -qi "git"; then
+                        git_issues="$git_issues\n    - $name: $install_hint"
+                    fi
+                    ;;
+                "shell"|"integration")
+                    shell_issues="$shell_issues\n    - $name: $install_hint"
+                    ;;
+            esac
+        fi
+    done
+    IFS="$old_ifs"
+    
+    # Priority-based recommendations
+    if [ -n "$critical_missing" ]; then
+        printf "\n${RED}🚨 CRITICAL ACTIONS REQUIRED:${NC}\n%s\n" "$critical_missing"
+    fi
+    
+    if [ -n "$recommended_missing" ]; then
+        printf "\n${YELLOW}📋 RECOMMENDED INSTALLATIONS:${NC}\n%s\n" "$recommended_missing"
+    fi
+    
+    # Category-specific advice
+    if [ -n "$git_issues" ]; then
+        printf "\n${BLUE}🔧 Git Configuration Issues:${NC}\n%s\n" "$git_issues"
+        printf "     ${GRAY}Tip: Configure git globally with user.name and user.email${NC}\n"
+    fi
+    
+    if [ -n "$shell_issues" ]; then
+        printf "\n${BLUE}🐚 Shell Configuration Issues:${NC}\n%s\n" "$shell_issues"
+        printf "     ${GRAY}Tip: Source shell files after installation or restart terminal${NC}\n"
+    fi
+}
+
+# Print chezmoi apply readiness status
+print_chezmoi_readiness_status() {
+    local adapter="$1"
+    local failed="$2" 
+    local score="$3"
+    
+    if [ $REPORT_LEVEL -ge $REPORT_NORMAL ]; then
+        printf "\n${BOLD}Chezmoi Apply Readiness${NC}\n"
+        printf "${GRAY}────────────────────────────────────────${NC}\n"
+        
+        # Analyze critical failures (required priority)
+        local critical_failures=0
+        local blocking_issues=""
+        
+        # Parse TEST_RESULTS for required failures
+        local old_ifs="$IFS"
+        IFS='
+'
+        for result_line in $TEST_RESULTS; do
+            IFS='|'
+            set -- $result_line
+            local category="$1"
+            local name="$2"
+            local test_status="$3"
+            local details="$4"
+            local priority="$5"
+            local install_hint="$6"
+            IFS="$old_ifs"
+            
+            if [ "$test_status" = "FAIL" ] && [ "$priority" = "required" ]; then
+                critical_failures=$((critical_failures + 1))
+                blocking_issues="$blocking_issues\n    🚫 $name ($category)"
+            fi
+        done
+        IFS="$old_ifs"
+        
+        # Determine readiness status
+        case "$adapter" in
+            pre_apply)
+                if [ $critical_failures -eq 0 ] && [ $score -ge 70 ]; then
+                    printf "${GREEN}✅ READY${NC} - Safe to run 'chezmoi apply'\n"
+                    printf "   ${GRAY}All critical requirements met (Health: ${score}%%)${NC}\n"
+                    
+                    if [ $failed -gt 0 ]; then
+                        printf "\n${YELLOW}⚠️  NON-BLOCKING ISSUES:${NC}\n"
+                        printf "   ${GRAY}Some optional/recommended components missing${NC}\n"
+                        printf "   ${GRAY}You can proceed but consider installing them later${NC}\n"
+                    fi
+                    
+                elif [ $critical_failures -eq 0 ] && [ $score -ge 50 ]; then
+                    printf "${YELLOW}⚠️  READY WITH WARNINGS${NC} - Can proceed with caution\n"
+                    printf "   ${GRAY}No critical failures but health is moderate (${score}%%)${NC}\n"
+                    printf "   ${GRAY}Consider fixing recommended issues first${NC}\n"
+                    
+                else
+                    printf "${RED}🚫 NOT READY${NC} - Do not run 'chezmoi apply' yet\n"
+                    printf "   ${GRAY}Critical failures must be resolved first${NC}\n"
+                    
+                    if [ -n "$blocking_issues" ]; then
+                        printf "\n${RED}🚨 BLOCKING ISSUES:${NC}%s\n" "$blocking_issues"
+                    fi
+                    
+                    printf "\n${YELLOW}NEXT STEPS:${NC}\n"
+                    printf "   1. Fix all critical (required) failures above\n" 
+                    printf "   2. Re-run test suite to verify fixes\n"
+                    printf "   3. Proceed with 'chezmoi apply' once ready\n"
+                fi
+                ;;
+                
+            post_apply)
+                if [ $critical_failures -eq 0 ] && [ $score -ge 80 ]; then
+                    printf "${GREEN}✅ OPTIMAL${NC} - Configuration is working well\n"
+                    printf "   ${GRAY}Environment is properly configured (Health: ${score}%%)${NC}\n"
+                    
+                elif [ $critical_failures -eq 0 ] && [ $score -ge 60 ]; then
+                    printf "${YELLOW}⚠️  FUNCTIONAL${NC} - Basic functionality working\n"
+                    printf "   ${GRAY}Core features available but some enhancements missing (${score}%%)${NC}\n"
+                    
+                else
+                    printf "${RED}🚫 NEEDS ATTENTION${NC} - Configuration issues detected\n"
+                    printf "   ${GRAY}Critical components missing or misconfigured${NC}\n"
+                    
+                    if [ -n "$blocking_issues" ]; then
+                        printf "\n${RED}🚨 CRITICAL ISSUES:${NC}%s\n" "$blocking_issues"
+                    fi
+                    
+                    printf "\n${YELLOW}NEXT STEPS:${NC}\n"
+                    printf "   1. Install missing critical components\n"
+                    printf "   2. Run 'chezmoi apply' to refresh configuration\n" 
+                    printf "   3. Restart shell sessions after fixes\n"
+                fi
+                ;;
+        esac
+        
+        # Additional context based on score
+        if [ $score -lt 30 ]; then
+            printf "\n${RED}💥 SEVERE ISSUES${NC}: Multiple critical components missing\n"
+        elif [ $score -lt 60 ]; then
+            printf "\n${YELLOW}⚠️  MODERATE ISSUES${NC}: Several components need attention\n"
+        elif [ $score -lt 90 ]; then
+            printf "\n${BLUE}ℹ️  MINOR ISSUES${NC}: Mostly good with some enhancements possible\n"
+        fi
     fi
 }
 
@@ -241,6 +430,7 @@ generate_report() {
     print_test_summary "$total" "$passed" "$failed" "$skipped"
     print_health_score "$total" "$passed" "$failed" "$skipped"
     local score=$?
+    print_chezmoi_readiness_status "$adapter" "$failed" "$score"
     print_recommendations "$failed" "$adapter"
     
     # Return exit code based on results