… dot_shell_common/CURRENT_STATUS.md
… dot_shell_common/INTEGRATION_PLAN.md
… dot_shell_common/dotfiles_doctor.sh.backup
… dot_shell_common/dotfiles_doctor_legacy.sh
… dot_shell_common/executable_doctor
… dot_shell_common/feature_parity_check.sh
… dot_shell_common/final_validation.sh